Adds dirs and moves files for tests
