+++ rpc/src/test/scala/org/bitcoins/rpc/common/MessageRpcTest.scala
@@ -1,4 +1,4 @@
-package org.bitcoins.rpc
+package org.bitcoins.rpc.common
 
 import akka.actor.ActorSystem
 import akka.stream.ActorMaterializer
@@ -6,6 +6,7 @@ import org.bitcoins.core.config.NetworkParameters
 import org.bitcoins.core.crypto.ECPrivateKey
 import org.bitcoins.core.protocol.P2PKHAddress
 import org.bitcoins.core.util.BitcoinSLogger
+import org.bitcoins.rpc.TestUtil
 import org.bitcoins.rpc.client.common.BitcoindRpcClient
 import org.bitcoins.rpc.client.common.RpcOpts.AddressType
 import org.scalatest.{ AsyncFlatSpec, BeforeAndAfter, BeforeAndAfterAll }
